refactor(routes): Implement Express Router for modular API structure
Refactors the API route handling logic using Express Router for better organization and maintainability.

- Introduces separate route files
- Defines API endpoints within their respective route files.
- Mounts the route files to the main Express app using `app.use()`.
- Improves code readability and separation of concerns.
